update divide and parameters test

+++ src/main/java/Calculator.java
@@ -37,10 +37,10 @@ public class Calculator {
         return a * b;
     }
 
-    public int Divide(int a, int b){
+    public double Divide(int a, int b){
         if (b == 0) {
             throw new IllegalArgumentException("Division by zero is not allowed.");
         }
-        return a / b;
+        return (double) a / b;
     }
 }
